Huge, painful overhaul of the IL with more detailed types
Some major cleanup is still necessary, but for now it compiles and the
tests pass. Huzzah.

+++ src/edu/umn/crisys/plexil/expr/common/LookupOnChangeExpr.java
@@ -16,7 +16,7 @@ public class LookupOnChangeExpr extends LookupExpr {
 	public LookupOnChangeExpr(ExprType type, Expression name, Expression tolerance, List<Expression> args) {
 		super(type, name, args);
 		
-	    ExprType.NUMERIC.typeCheck(tolerance.getType());
+	    ExprType.REAL.typeCheck(tolerance.getType());
 	    this.tolerance = tolerance;
 	}
